# docs/swagger.yaml
definitions:
...
  handler.CreateDepartureRequest:
    properties:
      ghi_chu:
        type: string
      ngay_ket_thuc:
        type: string
      ngay_khoi_hanh:
        type: string
      suc_chua:
        type: integer
      tour_id:
        type: integer
      trang_thai:
        type: string
    required:
    - ngay_ket_thuc
    - ngay_khoi_hanh
    - suc_chua
    - tour_id
    type: object
...
  handler.UpdateDepartureRequest:
    properties:
      ghi_chu:
        type: string
      ngay_ket_thuc:
        type: string
      ngay_khoi_hanh:
        type: string
      suc_chua:
        type: integer
      trang_thai:
        type: string
    type: object
  models.AddHinhAnhTourRequest:
    properties:
      duong_dan:
        type: string
      la_anh_chinh:
        type: boolean
      mo_ta:
        type: string
      thu_tu_hien_thi:
        type: integer
    required:
    - duong_dan
    type: object
...
  models.AddTourDestinationRequest:
    properties:
      diem_den_id:
        type: integer
      thu_tu_tham_quan:
        type: integer
    required:
    - diem_den_id
    type: object
...
  models.UpdateHoatDongTrongNgayRequest:
    properties:
      gio_bat_dau:
        description: 'Format: "HH:MM:SS"'
        type: string
      gio_ket_thuc:
        description: 'Format: "HH:MM:SS"'
        type: string
      mo_ta:
        type: string
      ten:
        type: string
      thu_tu:
        type: integer
    required:
    - ten
    type: object
  models.UpdateLichTrinhRequest:
    properties:
      dia_diem:
        type: string
      gio_bat_dau:
        description: 'Format: "HH:MM:SS"'
        type: string
      gio_ket_thuc:
        description: 'Format: "HH:MM:SS"'
        type: string
      mo_ta:
        type: string
      ngay_thu:
        type: integer
      thong_tin_luu_tru:
        type: string
      tieu_de:
        type: string
    required:
    - ngay_thu
    - tieu_de
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Travia API for travel management
  termsOfService: http://swagger.io/terms/
  title: Travia API
  version: "1.0"
paths:
...
  /departure/{id}:
    delete:
      description: Xóa một lịch khởi hành
      parameters:
      - description: Departure ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Xóa lịch khởi hành
      tags:
      - departure
    get:
      description: Lấy thông tin chi tiết của một lịch khởi hành
      parameters:
      - description: Departure ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Lấy chi tiết lịch khởi hành
      tags:
      - departure
    put:
      consumes:
      - application/json
      description: Cập nhật thông tin lịch khởi hành
      parameters:
      - description: Departure ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateDepartureRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cập nhật lịch khởi hành
      tags:
      - departure
  /departure/{id}/cancel:
    put:
      description: Hủy một lịch khởi hành (đặt status = huy)
      parameters:
      - description: Departure ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Hủy lịch khởi hành
      tags:
      - departure
  /departure/add-hinh-anh/{id}:
    post:
      consumes:
      - application/json
      description: Thêm hình ảnh tour
      parameters:
      - description: Tour ID
        in: path
        name: id
        required: true
        type: integer
      - description: Add data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AddHinhAnhTourRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Thêm hình ảnh tour
      tags:
      - departure
  /departure/add-tour-destination/{id}:
    post:
      consumes:
      - application/json
      description: Thêm một điểm đến tour
      parameters:
      - description: Tour ID
        in: path
        name: id
        required: true
        type: integer
      - description: Add data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AddTourDestinationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Thêm điểm đến tour
      tags:
      - departure
  /departure/create:
    post:
      consumes:
      - application/json
      description: Tạo lịch khởi hành mới cho tour
      parameters:
      - description: Departure data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateDepartureRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Tạo lịch khởi hành
      tags:
      - departure
  /departure/delete-hinh-anh/{id}:
    delete:
      description: Xóa một hình ảnh tour (cả trong database và Supabase storage)
      parameters:
      - description: Hình ảnh tour ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Xóa hình ảnh tour
      tags:
      - departure
  /departure/delete-tour-destination/{tour_id}/{diem_den_id}:
    delete:
      description: Xóa một điểm đến tour
      parameters:
      - description: Tour ID
        in: path
        name: tour_id
        required: true
        type: integer
      - description: Diem den ID
        in: path
        name: diem_den_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Xóa điểm đến tour
      tags:
      - departure
  /departure/hoat-dong-trong-ngay/{id}:
    put:
      consumes:
      - application/json
      description: Cập nhật hoạt động trong ngày
      parameters:
      - description: Hoạt động trong ngày ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateHoatDongTrongNgayRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cập nhật hoạt động trong ngày
      tags:
      - departure
  /departure/lich-trinh/{id}:
    put:
      consumes:
      - application/json
      description: Cập nhật lịch trình
      parameters:
      - description: Lịch trình ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateLichTrinhRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cập nhật lịch trình
      tags:
      - departure
  /departure/tour/{tour_id}:
    get:
      description: Lấy tất cả lịch khởi hành của một tour
      parameters:
      - description: Tour ID
        in: path
        name: tour_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Lấy lịch khởi hành theo tour
      tags:
      - departure
  /departure/upcoming:
    get:
      description: Lấy danh sách lịch khởi hành trong 30 ngày tới
      parameters:
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Lấy lịch khởi hành sắp tới
      tags:
      - departure
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"